Reorganize Hibernate Search management config
While it's working fine with the new annotation processor, the config is
not ordered the same and it's better displayed as a section.
Also it looks a bit odd to have a specific root just for that.
