+++ extensions/hibernate-search-orm-elasticsearch/runtime/src/main/java/io/quarkus/hibernate/search/orm/elasticsearch/runtime/HibernateSearchElasticsearchBuildTimeConfigManagement.java
@@ -1,13 +1,10 @@
-package io.quarkus.hibernate.search.orm.elasticsearch.runtime.management;
+package io.quarkus.hibernate.search.orm.elasticsearch.runtime;
 
-import io.quarkus.runtime.annotations.ConfigPhase;
-import io.quarkus.runtime.annotations.ConfigRoot;
-import io.smallrye.config.ConfigMapping;
+import io.quarkus.runtime.annotations.ConfigGroup;
 import io.smallrye.config.WithDefault;
 
-@ConfigMapping(prefix = "quarkus.hibernate-search-orm.management")
-@ConfigRoot(phase = ConfigPhase.BUILD_AND_RUN_TIME_FIXED)
-public interface HibernateSearchManagementConfig {
+@ConfigGroup
+public interface HibernateSearchElasticsearchBuildTimeConfigManagement {
 
     /**
      * Root path for reindexing endpoints.
